Extracted classes from main script into separate files
