+++ clients/influxdbdata09.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+class InfluxDBData09(object):
+	def __init__(self):
+		self.points = []
+
+	def add_points(self, points):
+		self.points = self.points + points
+
+	def reset(self):
+		self.points = []
